stock/predict/requirements.txt
pytube3
moviepy
SpeechRecognition
pydub
ffmpeg-python
